requirements.txt
flask==2.0.2
gunicorn==20.1.0
wordninja==2.0.0
nltk==3.8.1
tensorflow-cpu==2.12.0
tensorflow-text==2.12.1
numpy==1.23.2
scipy==1.10.1
scikit-learn==1.2.2
